✨: Add a new Flutter project named 'test'
This commit introduces a new Flutter project.  The project is created using the brick template and is named 'test'.  The project is configured to not be published.  The brick.yaml file is updated to remove the name variable.  The .dart_tool and .flutter-plugins directories are generated as part of the project creation.